build: compile as UTF-8 on MSVC, and spell out the narrowings
The new Windows job failed on the first translation unit it compiled:

    cats.hpp(28): error C2220: the following warning is treated as an error
    cats.hpp(28): warning C4566: character represented by
        universal-character-name '\u03C9' cannot be represented in the
        current code page (1252)

MSVC defaults both character sets to the machine code page. Every cat face
and every box drawing character is multi-byte, so `\u...` in a string
literal is unrepresentable — an error under /WX — and the same literals
would have reached the terminal mangled even without it. `/utf-8` states
what the project already is: UTF-8 in, UTF-8 out. It is set before any
subdirectory is added so the fetched dependencies compile the same way.

`terminal_win32.cpp` gets explicit `DWORD` casts where a `size_t` was
narrowed implicitly, which /W4 reports as C4267 and /WX turns into an error.

+++ src/tui/terminal_win32.cpp
@@ -167,7 +167,7 @@ private:
     for (;;) {
       char buf[256];
       DWORD n = 0;
-      if (!::ReadFile(in_, buf, sizeof(buf), &n, nullptr)) {
+      if (!::ReadFile(in_, buf, static_cast<DWORD>(sizeof(buf)), &n, nullptr)) {
         break; // a dying console is not EOF; just stop
       }
       if (n == 0) {
@@ -196,7 +196,7 @@ private:
       if (!::PeekNamedPipe(in_, nullptr, 0, nullptr, &avail, nullptr)) {
         char buf[256];
         DWORD n = 0;
-        if (!::ReadFile(in_, buf, sizeof(buf), &n, nullptr) || n == 0) {
+        if (!::ReadFile(in_, buf, static_cast<DWORD>(sizeof(buf)), &n, nullptr) || n == 0) {
           closed_ = true; // broken pipe, or a file at EOF
         } else {
           out.append(buf, n);
@@ -206,7 +206,7 @@ private:
       if (avail > 0) {
         char buf[256];
         DWORD n = 0;
-        const DWORD want = static_cast<DWORD>(std::min<DWORD>(avail, sizeof(buf)));
+        const DWORD want = std::min<DWORD>(avail, static_cast<DWORD>(sizeof(buf)));
         if (::ReadFile(in_, buf, want, &n, nullptr) && n > 0) {
           out.append(buf, n);
         }
